TCOMP-200 : make Properties an interface - temporary commit
