+++ daikon/src/main/java/org/talend/daikon/properties/PropertiesDynamicMethodHelper.java
@@ -62,9 +62,9 @@ public class PropertiesDynamicMethodHelper {
 
     public static void storeResult(Properties props, Object result) {
         if (result instanceof ValidationResult && result != null) {
-            props.validationResult = (ValidationResult) result;
+            props.setValidationResult((ValidationResult) result);
         } else {
-            props.validationResult = ValidationResult.OK;
+            props.setValidationResult(ValidationResult.OK);
         }
     }
 
@@ -83,7 +83,7 @@ public class PropertiesDynamicMethodHelper {
         Method m = findMethod(props, Properties.METHOD_VALIDATE, propName, REQUIRED);
         try {
             m.setAccessible(true);
-            props.validationResult = (ValidationResult) m.invoke(props);
+            props.setValidationResult((ValidationResult) m.invoke(props));
         } catch (InvocationTargetException e) {
             throw e.getTargetException();
         }
