+++ daikon/src/test/java/org/talend/daikon/properties/presentation/FormTest.java
@@ -17,13 +17,13 @@ import static org.talend.daikon.properties.presentation.Widget.*;
 import static org.talend.daikon.properties.property.PropertyFactory.*;
 
 import org.junit.Test;
-import org.talend.daikon.properties.Properties;
+import org.talend.daikon.properties.PropertiesImpl;
 
 public class FormTest {
 
     @Test
     public void testNameEqualsDisplayNameWithUsualConstructor() {
-        Form form = new Form(new Properties("bar") { //$NON-NLS-1$
+        Form form = new Form(new PropertiesImpl("bar") { //$NON-NLS-1$
         }, "foo"); //$NON-NLS-1$
         assertEquals("foo", form.getName()); //$NON-NLS-1$
         assertEquals("foo", form.getDisplayName()); //$NON-NLS-1$
@@ -31,7 +31,7 @@ public class FormTest {
 
     @Test
     public void testGetI18NFields() {
-        Form form = new Form(new Properties("bar") { //$NON-NLS-1$
+        Form form = new Form(new PropertiesImpl("bar") { //$NON-NLS-1$
         }, "foo", null); //$NON-NLS-1$
         assertEquals("Ze Form DisplayName", form.getDisplayName()); //$NON-NLS-1$
         assertEquals("Ze Form Title", form.getTitle()); //$NON-NLS-1$
@@ -42,7 +42,7 @@ public class FormTest {
     public void testGetI18NFiledsWithDefaultValue() {
         String displayName = "Default Display Name"; //$NON-NLS-1$
         String subTitle = "Default SubTitle"; //$NON-NLS-1$
-        Form form = new Form(new Properties("bar") { //$NON-NLS-1$
+        Form form = new Form(new PropertiesImpl("bar") { //$NON-NLS-1$
         }, "foo", displayName); //$NON-NLS-1$
         form.setSubtitle(subTitle);
         assertEquals(displayName, form.getDisplayName());
@@ -52,7 +52,7 @@ public class FormTest {
 
     @Test
     public void testSetVisible() {
-        Form form = new Form(new Properties("bar") { //$NON-NLS-1$
+        Form form = new Form(new PropertiesImpl("bar") { //$NON-NLS-1$
         }, "foo"); //$NON-NLS-1$
         form.addRow(widget(newString("w1")));
         form.addRow(widget(newString("w2")));
@@ -72,9 +72,9 @@ public class FormTest {
 
     @Test
     public void testSetVisibleForNestedForms() {
-        Form form = new Form(new Properties("bar") { //$NON-NLS-1$
+        Form form = new Form(new PropertiesImpl("bar") { //$NON-NLS-1$
         }, "foo"); //$NON-NLS-1$
-        Form nestedForm = new Form(new Properties("foo") { //$NON-NLS-1$
+        Form nestedForm = new Form(new PropertiesImpl("foo") { //$NON-NLS-1$
         }, "bar"); //$NON-NLS-1$
         form.addRow(widget(newString("w1")));
         form.addRow(widget(nestedForm));
