+++ daikon/src/test/java/org/talend/daikon/properties/testproperties/TestProperties.java
@@ -24,6 +24,7 @@ import org.apache.avro.Schema;
 import org.apache.avro.SchemaBuilder;
 import org.talend.daikon.properties.PresentationItem;
 import org.talend.daikon.properties.Properties;
+import org.talend.daikon.properties.PropertiesImpl;
 import org.talend.daikon.properties.ValidationResult;
 import org.talend.daikon.properties.ValidationResult.Result;
 import org.talend.daikon.properties.presentation.Form;
@@ -33,7 +34,7 @@ import org.talend.daikon.properties.service.Repository;
 import org.talend.daikon.properties.testproperties.nestedprop.NestedProperties;
 import org.talend.daikon.properties.testproperties.nestedprop.inherited.InheritedProperties;
 
-public class TestProperties extends Properties {
+public class TestProperties extends PropertiesImpl {
 
     public static final String USER_ID_PROP_NAME = "userId"; //$NON-NLS-1$
